database: organization and tests cleanups
The structure (and most of the code) was taken from revaultd. This was
not appropriate to separate it into the 'write' and 'read' sections as
in revaultd since our implementation is pretty straightforward: two
setters, two getters and the db creation routine.
